requirements.txt
Flask
flask-cors
pymongo
shortuuid
python-dotenv
